Revert "Use float instead of double"
This reverts commit 25dce4a9bfc28159439f532b2ac86cd25ec5d59f.

+++ cpp/Sample.h
@@ -11,12 +11,12 @@ public:
 
     bool run();
 
-    float r2;
+    double r2;
 };
 
 inline bool Sample::run() {
     for (int j = 0; j < d; j++) {
-        float x_j = get_random_0_to_1();
+        double x_j = get_random_0_to_1();
         r2 += pow(x_j, 2);
         if (r2 > 1) return false;
     }
